feat: apply styles to form, chat page and stuff
