+++ client/src/App.css
@@ -1,67 +1,5 @@
 @import "https://cdn.jsdelivr.net/npm/bulma@0.9.4/css/bulma-rtl.min.css";
 
-.bulma-center-mixin {
-  @include center;
-}
-
 .card{
   width: 15vw;
 }
-
-table {
-  width: 200px;
-  height: 200px;
-}
-
-th {
-  height: 20%;
-  background-color: #dce5f9;
-}
-
-#add-button {
-  margin-left: 6px;
-}
-
-#home-buttons-container {
-  display: flex;
-}
-
-#home-buttons-container div {
-  margin: 1px;
-}
-
-.tutor-introduction {
-  border: 1px dashed black;
-  height: 300px;
-  width: 80%;
-  background-color: rgb(254, 254, 254);
-}
-
-ul {
-  list-style-type: none;
-  display: flex;
-}
-
-ul li {
-  text-align: center;
-}
-
-ul li a {
-  color: #000;
-  margin: 10px;
-  text-align: center;
- }
-
- ul li a:hover {
-  color: #00a8ff;
- }
-
- ul li a.active {
-  color: #1900ff;
- }
-
-#tutor-registration-form {
-  display: flex;
-  flex-direction: column;
- }
-
