+++ client/src/Pages/not-found-page.tsx
@@ -1,10 +1,8 @@
 import React from 'react';
-import NavBar from '../components/nav-bar';
 
 function NotFound () {
   return (
     <>
-      <NavBar />
       <h1 className='title'>Content Not Found</h1>
     </>
   );
