+++ client/src/Pages/tutors-list.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import TutorInterface from '../custom-types/types';
 import TutorCard from '../components/tutor-card';
 import { useEffect } from 'react';
-import NavBar from '../components/nav-bar';
 
 type Props = {
   tutors: TutorInterface[]
@@ -14,7 +13,6 @@ function TutorsList({ tutors }: Props) {
 
   return (
     <>
-      <NavBar />
       <div className='container is-flex is-flex-direction-row is-flex-wrap-wrap'>
       {
           tutors.map((tutor) => {
