+++ client/src/components/sign-up.tsx
@@ -25,10 +25,10 @@ export default function SignUpModal({ toggleSignUpModal }: Props) {
           <p className="subtitle">Are you a tutor or a student?</p>
         </section>
         <footer className="modal-card-foot is-flex is-justify-content-center">
-          <button className="button is-success ml-3 mr-3">
+          <button className="button is-success ml-3 mr-3"  onClick={(toggleSignUpModal)}>
             <NavLink to='/register'>Tutor</NavLink>
           </button>
-          <button className="button is-warning ml-3 mr-3">
+          <button className="button is-warning ml-3 mr-3" onClick={(toggleSignUpModal)}>
             <NavLink to='/studentsignup'>
               Student
             </NavLink>
